Refactor ShardingSphereInsideException
