+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/exception/standard/SQLStandardException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.error.exception.standard;
 
-import org.apache.shardingsphere.infra.util.exception.inside.ShardingSphereInsideException;
+import org.apache.shardingsphere.infra.util.exception.ShardingSphereInsideException;
 
 /**
  * SQL standard exception.
