+++ shardingsphere-infra/shardingsphere-infra-util/src/main/java/org/apache/shardingsphere/infra/util/exception/ShardingSphereInsideException.java
@@ -15,23 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.util.exception.inside.fixture;
+package org.apache.shardingsphere.infra.util.exception;
 
-import org.apache.shardingsphere.infra.util.exception.inside.ShardingSphereInsideException;
+import lombok.NoArgsConstructor;
 
-public final class ShardingSphereInsideExceptionFixture extends ShardingSphereInsideException {
+/**
+ * ShardingSphere inside exception.
+ */
+@NoArgsConstructor
+public abstract class ShardingSphereInsideException extends RuntimeException {
     
-    private static final long serialVersionUID = 3759417279204541147L;
+    private static final long serialVersionUID = -8238061892944243621L;
     
-    public ShardingSphereInsideExceptionFixture(final String message) {
-        super("Fixture error message: %s", message);
+    public ShardingSphereInsideException(final String errorMessage, final Object... args) {
+        super(String.format(errorMessage, args));
     }
     
-    public ShardingSphereInsideExceptionFixture(final String message, final Exception cause) {
-        super(message, cause);
-    }
-    
-    public ShardingSphereInsideExceptionFixture(final Exception cause) {
+    public ShardingSphereInsideException(final Exception cause) {
         super(cause);
+    }
+    
+    public ShardingSphereInsideException(final String message, final Exception cause) {
+        super(message, cause);
     }
 }
