+++ shardingsphere-infra/shardingsphere-infra-util/src/test/java/org/apache/shardingsphere/infra/util/exception/fixture/ShardingSphereInsideExceptionFixture.java
@@ -15,27 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.util.exception.inside;
+package org.apache.shardingsphere.infra.util.exception.fixture;
 
-import lombok.NoArgsConstructor;
+import org.apache.shardingsphere.infra.util.exception.ShardingSphereInsideException;
 
-/**
- * ShardingSphere inside exception.
- */
-@NoArgsConstructor
-public abstract class ShardingSphereInsideException extends RuntimeException {
+public final class ShardingSphereInsideExceptionFixture extends ShardingSphereInsideException {
     
-    private static final long serialVersionUID = -8238061892944243621L;
+    private static final long serialVersionUID = 3759417279204541147L;
     
-    public ShardingSphereInsideException(final String errorMessage, final Object... args) {
-        super(String.format(errorMessage, args));
+    public ShardingSphereInsideExceptionFixture(final String message) {
+        super("Fixture error message: %s", message);
     }
     
-    public ShardingSphereInsideException(final Exception cause) {
-        super(cause);
-    }
-    
-    public ShardingSphereInsideException(final String message, final Exception cause) {
+    public ShardingSphereInsideExceptionFixture(final String message, final Exception cause) {
         super(message, cause);
+    }
+    
+    public ShardingSphereInsideExceptionFixture(final Exception cause) {
+        super(cause);
     }
 }
